+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add `LWBTN_CFG_GET_STATE_MODE` to control *get state* mode
 - Add option to check if button is currently active (after debounce period has elapsed) 
+- Add option to set time/click parameters at run time for each button specifically
 
 ## v0.0.1